fix logging again

+++ src/main.cu
@@ -44,13 +44,13 @@ int main(int argc, const char ** argv) {
             step_log_cfg.agent("Neuroblastoma").logMean<int>("necro");
             step_log_cfg.agent("Neuroblastoma").logMean<int>("necro_signal");
             step_log_cfg.agent("Neuroblastoma").logMean<int>("apop_signal");
-            step_log_cfg.agent("Neuroblastoma").logMean<float>("cycle");
+            step_log_cfg.agent("Neuroblastoma").logMean<unsigned int>("cycle");
             step_log_cfg.agent("Schwann").logCount();
             step_log_cfg.agent("Schwann").logMean<int>("apop");
             step_log_cfg.agent("Schwann").logMean<int>("necro");
             step_log_cfg.agent("Schwann").logMean<int>("necro_signal");
             step_log_cfg.agent("Schwann").logMean<int>("apop_signal");
-            step_log_cfg.agent("Schwann").logMean<float>("cycle");
+            step_log_cfg.agent("Schwann").logMean<unsigned int>("cycle");
         }
         /**
          * Create Model Runner
